requirements.txt
line-bot-sdk
flask
google-genai
python-dotenv
gunicorn
